# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Portal Manajemen Proyek API
  version: 1.0.0
  description: REST API untuk autentikasi, RBAC, CRUD proyek, tugas, dashboard, risiko, time tracking, file, dan komentar.
servers:
  - url: http://localhost:4000/api
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: Daftar proyek sesuai role user login
      responses:
        "200":
          description: Daftar proyek terfilter RBAC
    post:
      tags: [Projects]
      summary: Membuat proyek, hanya Project Manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Proyek dibuat
        "403":
          description: Role tidak diizinkan
  /projects/{id}:
    put:
      tags: [Projects]
      summary: Mengubah proyek milik Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Proyek diperbarui
    delete:
      tags: [Projects]
      summary: Menghapus proyek milik Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Proyek dihapus
  /tasks:
    get:
      tags: [Tasks]
      summary: Daftar tugas sesuai role, status, projectId, dan rentang tanggal
      parameters:
        - in: query
          name: projectId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [todo, in_progress, done]
        - in: query
          name: fromDate
          schema:
            type: string
            format: date
        - in: query
          name: toDate
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daftar tugas terfilter
    post:
      tags: [Tasks]
      summary: Membuat tugas, hanya Project Manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Tugas dibuat
  /tasks/{id}:
    put:
      tags: [Tasks]
      summary: Project Manager mengubah tugas, Developer mengubah status/progress tugasnya
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "200":
          description: Tugas diperbarui
    delete:
      tags: [Tasks]
      summary: Menghapus tugas, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Tugas dihapus
...
  /risks/{id}:
    put:
      tags: [Dashboard]
      summary: Mengubah risiko, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Risiko diperbarui
    delete:
      tags: [Dashboard]
      summary: Menghapus risiko, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Risiko dihapus
...
  /task-comments/{id}:
    put:
      tags: [Dashboard]
      summary: Mengubah komentar oleh pembuat komentar atau Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Komentar diperbarui
    delete:
      tags: [Dashboard]
      summary: Menghapus komentar, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Komentar dihapus
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        status:
          type: string
          enum: [planning, in_progress, completed, on_hold, on_track, at_risk, delayed]
        client_id:
          type: integer
        pm_id:
          type: integer
        cover_image_url:
          type: string
          format: uri
    TaskInput:
      type: object
      required: [project_id, name]
      properties:
        project_id:
          type: integer
        name:
          type: string
        description:
          type: string
        assigned_to:
          type: integer
        status:
          type: string
          enum: [todo, in_progress, done]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        due_date:
          type: string
          format: date